% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrdiscdist-package.R
\docType{package}
\name{corrdiscdist-package}
\alias{corrdiscdist-package}
\alias{corrdiscdist}
\title{What the Package Does (One Line, Title Case)}
\description{
Implementation of simulation (wrapping around RcppDist) and inference (wrapping around mvtnorm) for selected correlated discrete distributions (negative binomial, binomial, ordinal multinomial). Provides interfaces for R, C++, JAGS and TMB (maybe).
}
\details{
Some description of the package
}
\examples{
## An example function:
df <- example_function()
df

## To cite this package in publications use:
citation("corrdiscdist")

}
\references{
Some article that you might want users to look at
}
\seealso{
\code{\link[neatpkg]{pkg_new}} for the function that created this package
}
\author{
\strong{Maintainer}: Matt Denwood \email{md@sund.ku.dk}

}
\keyword{internal}
